# backend/swagger.yaml
openapi: 3.0.0
info:
  title: GigFlow API
  description: |
    Full-stack freelance marketplace API where Clients can post jobs (Gigs) and Freelancers can apply for them (Bids).
    
    ## Features
    - Secure authentication with JWT tokens in HttpOnly cookies
    - Gig management (CRUD operations)
    - Bidding system
    - Hiring logic with transactional integrity
    - Real-time notifications via Socket.io
    
    ## Authentication
    All protected endpoints require a valid JWT token stored in an HttpOnly cookie.
    The token is automatically set when you register or login.
  version: 1.0.0
  contact:
    name: GigFlow API Support
    email: support@gigflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://api.gigflow.com/api
    description: Production server
...
paths:
...
  /gigs:
    get:
      tags:
        - Gigs
      summary: Get all gigs
      description: Retrieves a list of gigs with optional search and status filtering
      operationId: getGigs
      parameters:
        - name: search
          in: query
          description: Search query to filter gigs by title or description
          required: false
          schema:
            type: string
            example: web development
        - name: status
          in: query
          description: Filter gigs by status
          required: false
          schema:
            type: string
            enum: [open, assigned]
            default: open
            example: open
      responses:
        '200':
          description: List of gigs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 10
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Gig'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Gigs
      summary: Create a new gig
      description: Creates a new job posting (requires authentication)
      operationId: createGig
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - budget
              properties:
                title:
                  type: string
                  maxLength: 100
                  example: Build a responsive e-commerce website
                description:
                  type: string
                  maxLength: 1000
                  example: Looking for an experienced web developer to build a modern e-commerce platform with React and Node.js
                budget:
                  type: number
                  minimum: 1
                  example: 5000
      responses:
        '201':
          description: Gig created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Gig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /gigs/{id}:
    get:
      tags:
        - Gigs
      summary: Get a single gig
      description: Retrieves detailed information about a specific gig
      operationId: getGig
      parameters:
        - name: id
          in: path
          required: true
          description: Gig ID
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Gig retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Gig'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Gigs
      summary: Update a gig
      description: Updates an existing gig (only the owner can update)
      operationId: updateGig
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Gig ID
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 1000
                budget:
                  type: number
                  minimum: 1
      responses:
        '200':
          description: Gig updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Gig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Gigs
      summary: Delete a gig
      description: Deletes a gig (only the owner can delete)
      operationId: deleteGig
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Gig ID
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Gig deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in HttpOnly cookie
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Manya Shukla
        email:
          type: string
          format: email
          example: shuklamanya99@gmail.com
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
...
    Gig:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          example: Build a responsive e-commerce website
        description:
          type: string
          example: Looking for an experienced web developer...
        budget:
          type: number
          example: 5000
        ownerId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
          example: 507f1f77bcf86cd799439011
        status:
          type: string
          enum: [open, assigned]
          example: open
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message describing what went wrong
  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation error message
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not authorized to access this route
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error